+++ examples/stackflow-spa/src/activities/ActivityHome.tsx
@@ -1,7 +1,6 @@
 import {
   Box,
   Divider,
-  Icon,
   Portal,
   PullToRefresh,
   VStack,
@@ -280,6 +279,7 @@ const ActivityHome: StaticActivityComponentType<"ActivityHome"> = ({ params }) =
         { title: "MannerTempLevel", onClick: () => push("ActivityMannerTempLevel", {}) },
         { title: "ErrorState", onClick: () => push("ActivityErrorState", {}) },
         { title: "ResultSection", onClick: () => push("ActivityResultSection", {}) },
+        { title: "SideNavigation", onClick: () => push("ActivitySideNavigation", {}) },
       ],
     },
     {