+++ examples/stackflow-spa/src/stackflow/Stack.tsx
@@ -82,6 +82,7 @@ export const { Stack, actions, stepActions } = stackflow({
     ActivityResultSection: lazy(() => import("../activities/ActivityResultSection")),
     ActivitySegmentedControl: lazy(() => import("../activities/ActivitySegmentedControl")),
     ActivitySwitch: lazy(() => import("../activities/ActivitySwitch")),
+    ActivitySideNavigation: lazy(() => import("../activities/ActivitySideNavigation")),
     ActivitySwipeableTabs: lazy(() => import("../activities/ActivitySwipeableTabs")),
     ActivityTabs: lazy(() => import("../activities/ActivityTabs")),
     ActivityForm: lazy(() => import("../activities/ActivityForm")),