+++ packages/qvism-preset/src/recipes.ts
@@ -55,6 +55,11 @@ import reactionButton from "./recipes/reaction-button";
 import scrollFog from "./recipes/scroll-fog";
 import segmentedControl from "./recipes/segmented-control";
 import { selectBox, selectBoxCheckmark, selectBoxGroup } from "./recipes/select-box";
+import {
+  sideNavigation,
+  sideNavigationInset,
+  sideNavigationMenuItem,
+} from "./recipes/side-navigation";
 import skeleton from "./recipes/skeleton";
 import { slider, sliderMarker, sliderTick } from "./recipes/slider";
 import { snackbar, snackbarRegion } from "./recipes/snackbar";
@@ -134,6 +139,9 @@ export const recipes = {
   selectBox,
   selectBoxCheckmark,
   selectBoxGroup,
+  sideNavigation,
+  sideNavigationInset,
+  sideNavigationMenuItem,
   skeleton,
   slider,
   sliderMarker,