+++ packages/react/src/components/index.ts
@@ -59,6 +59,7 @@ export * from "./ResponsivePair";
 export * from "./ScrollFog";
 export * from "./SegmentedControl";
 export * from "./SelectBox";
+export * from "./SideNavigation";
 export * from "./Skeleton";
 export * from "./Slider";
 export * from "./Snackbar";