+++ packages/react-headless/tabs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @seed-design/react-tabs
 
+## 0.0.0-alpha-20260414104312
+
+### Patch Changes
+
+- Updated dependencies
+  - @seed-design/react-primitive@0.0.0-alpha-20260414104312
+
 ## 1.0.3
 
 ### Patch Changes